BAVL-1084 fix minor CVP details validation
This is a minor issue but would be nice to fix.

During editing of a court booking with a CVP link or HMCTS number and hit continue it repopulates the fields you cleared out with those from the session but still shows the error please provide a link.

This should not happen, the fields should remain blank.